refactor(utility): moved out logic for  and some graph helpers into utility.py
